fix: 修复部署bug
